skip adding apt keyring if already installed

+++ installer/ubuntu2404_clang19.bash
@@ -5,8 +5,10 @@ set -euo pipefail
 LLVM_VERSION=19
 UBUNTU_CODENAME=noble
 
-mkdir -p /usr/local/share/keyrings
-curl --tlsv1.2 -sSf https://apt.llvm.org/llvm-snapshot.gpg.key | gpg --dearmor -o /usr/local/share/keyrings/llvm-snapshot-archive-keyring.gpg
+if [ ! -f /usr/local/share/keyrings/llvm-snapshot-archive-keyring.gpg ]; then
+  mkdir -p /usr/local/share/keyrings
+  curl --tlsv1.2 -sSf https://apt.llvm.org/llvm-snapshot.gpg.key | gpg --dearmor -o /usr/local/share/keyrings/llvm-snapshot-archive-keyring.gpg
+fi
 echo "deb [signed-by=/usr/local/share/keyrings/llvm-snapshot-archive-keyring.gpg] http://apt.llvm.org/${UBUNTU_CODENAME}/ llvm-toolchain-${UBUNTU_CODENAME}-${LLVM_VERSION} main" \
   >> /etc/apt/sources.list.d/llvm.list
 echo "deb-src [signed-by=/usr/local/share/keyrings/llvm-snapshot-archive-keyring.gpg] http://apt.llvm.org/${UBUNTU_CODENAME}/ llvm-toolchain-${UBUNTU_CODENAME}-${LLVM_VERSION} main" \
